+++ x/posts/keeper/keeper_test.go
@@ -1,8 +1,9 @@
 package keeper_test
 
 import (
-	"github.com/desmos-labs/desmos/x/posts/types/models/common"
 	"time"
+
+	"github.com/desmos-labs/desmos/x/posts/types/models/common"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"