+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 - Removed the user_block event attribute key prefix (#291)
 - Changed posts' optional data representation (#272)
 - Fixed bugs inside the DTag transfer process (#295, #296)
+- Implemented the possibility to refuse and cancel DTag requests from both receiver and sender side (#297)
 
 ## Bug fixes
 - Fixed the possibility of requesting a transfer of an empty DTag (#292)
+- Fixed the impossibility  of querying all the relationships on chain (#306)
 
 # Version 0.12.3
 ## Changes